+++ packages/golden-layout/changelog.md
@@ -1,3 +1,5 @@
+2.3.0
+feat: added support for maximization boundaries
 2.2.8
 fix: isMaximized property wasn't valid after adding a new item while there is a maximized item
 2.2.7